ingest/requirements.txt
-i https://pypi.python.org/simple
adal==1.2.2
antlr4-python3-runtime==4.8 ; python_version >= '3.0'
applicationinsights==0.11.9
argcomplete==1.11.1
azure-cli-core==2.0.81
azure-cli-nspkg==3.0.4
azure-cli-telemetry==1.0.4
azure-common==1.1.24
azure-core==1.2.1
azure-mgmt-billing==0.2.0
azure-mgmt-consumption==3.0.0
azure-mgmt-datafactory==0.8.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==6.0.0
azure-nspkg==3.0.2
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
humanfriendly==4.18
idna==2.8
importlib-metadata==1.5.0 ; python_version == '3.5'
isodate==0.6.0
jmespath==0.9.4
knack==0.6.3
msrest==0.6.11
msrestazure==0.6.2
oauthlib==3.1.0
paramiko==2.7.1
portalocker==1.5.2
pycparser==2.19
pygments==2.5.2
pyjwt==1.7.1
pynacl==1.3.0
pyopenssl==19.1.0
python-dateutil==2.8.1
pyyaml==5.3
requests-oauthlib==1.3.0
requests==2.22.0
six==1.14.0
tabulate==0.8.6
urllib3==1.25.8
wheel==0.30.0
zipp==1.1.0
